hal(spi): widen SpiHal::SetClockDiv div from uint16_t to uint32_t

+++ hal/include/backbone/spi_hal.h
@@ -65,7 +65,7 @@ class SpiHal {
      * @param div Even divider from the peripheral clock.
      * @return HalStatus::Ok on success.
      */
-    HalStatus SetClockDiv(uint16_t div);
+    HalStatus SetClockDiv(uint32_t div);
 
     /**
      * @brief Manually drive a chip-select line.
